+++ CHANGELOG.md
@@ -29,9 +29,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - **Breaking**: move the `zarrs::array::{data_type,fill_value}` modules into the `zarrs_data_type` crate
 - Bump `lru` to 0.13
 - Use codec identifiers in the example for `experimental_codec_names` remapping
-
-### Fixed
-- Error on `{Array,Group}::[async_]open[_opt]` with additional fields with `"must_understand": true`
+- Allow `{Array,Group}::new_with_metadata()` and `{Array,Group}Builder` to create arrays with `"must_understand": true` additional fields
+  - `{Array,Group}::[async_]open[_opt]` continue to fail with additional fields with `"must_understand": true`
 
 ## [0.19.2] - 2025-02-13